Add CVMix to Externals_MOM, don't build MARBL
To take CVMix and MARBL out of the CESM Externals.cfg, buildlib will now look
for pkg/CVMix-src (which MOM6 knows about via submodules). Since MARBL is not
yet a submodule, it is omitted from the build system entirely - I'll add it
back in when I issue a PR for my branch that includes a MARBL driver.