Group calls needed to add MARBL
Eventually MARBL will only be built if a BGC-enabled compset is specified, so
it makes sense to only look for pkg/MARBL and add it to paths if it will be
needed. At this point it is in an "if True:" block, but down the line the
condition will change.