3D tracers should be on z levels
This means splitting all hm_bgc streams into hm_bgc_{freq} and hm_bgc_{freq}_z,
with the 2D fields in the former and 3D fields in the latter. It also means
generate_diag_table_MARBL() needs to know which fields are 2D and which are 3D;
for MARBL fields it can look at the vertical_grid key in the diagnostics JSON
file but for the MOM6 fields we assume STF_{tracer} are the only 2D fields for
now.