Make sure MARBL streams are archived / tested
Previously, ERS tests weren't comparing the hm_bgc streams that MOM was
producing - this let a bug slip in where river fluxes weren't set up properly
on restart runs. I updated the logic in config_archive.xml to make sure h_bgc
and hm_bgc streams are tested, and then I updated the diag_table so that the
test writes hm_bgc_YYYY_MM (so a 5-day test will create this file when writing
output on day one and then over-write the data each subsequent day). This is
necessary because otherwise a 5-day test will generate a tremendous amount of
output for cprnc to check.