Easier to switch between monthly and nstep output
For the h[m]_bgc (and _z) streams, there's now a single logical flag
controlling whether this output is monthly (default) or nstep (useful for
debugging). If this is a permanent feature then it needs to be implemented in a
cleaner manner, perhaps with an env_run.xml flag, but it could be that this
gets pulled out once initial testing is complete.