Introduction of handling unresolved diag_tables:
- The buildnml script now generates a "diag_table.unresolved" file alongside
the usual "diag_table" in the Buildconf/momconf directory. In this unresolved
 version, the case name is represented as ${CASE}, enabling easy sharing of a
 diag table across multiple cases in a portable manner. This unresolved diag
table can be placed in SourceMods/src.mom and modified if needed. If a
diag_table.uresolved file is found in SourceMods/src.mom, the buildnml script
creates a new "diag_table" file in the RUNDIR directory based on the unresolved
one, replacing ${CASE} with  the actual case name. If both diag_table and
diag_table.unresolved files are encountered in the SourceMods directory, an
error will be triggered to instruct the user to use only one of them.