Update to marbl0.47.0
This MARBL tag introduces the idea of diag_mode to help select which
diagnostics are included in the output. Support in MOM requires a new
env_run.xml variable (MARBL_DIAG_MODE), which gets passed from buildnml to
several of the MARBL_scripts.

Note that the MARBL diagnostics are now output in single precision unless
MARBL_DIAG_MODE == "test_suite" (in which case it gets bumped up to double
precision).