Clean up previous commit. Please see below for details
To run MOM standalone use ./build_examples_cesm.sh just like
./build_examples.sh. To use the CESM compiler flags, (or at least the
rough conversion I've done), add a "--cesm" arg to the run. The
NVHPC with CESM is not working, and with intel debug it is clear there
is an issue. The GNU and Intel CESM version (Non Debug) builds and
runs. See templates for the changes between non-cesm and cesm flags.

If you would like to use the builds without building yourself, check out
this repo on derecho in
/glade/u/home/manishrv/documents/installs/mom_interface_pr_192/components/mom/standalone/build.